+++ nion/data/xdata_1_0.py
@@ -249,11 +249,17 @@ def invert(data_and_metadata: DataAndMetadata.DataAndMetadata) -> DataAndMetadat
 def register_translation(xdata1: DataAndMetadata.DataAndMetadata, xdata2: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, subtract_means: bool = True) -> typing.Tuple[float, ...]:
     return Core.function_register(xdata1, xdata2, upsample_factor, subtract_means)
 
-def shift(src: DataAndMetadata.DataAndMetadata, shift: typing.Tuple[float, ...]) -> DataAndMetadata.DataAndMetadata:
-    return Core.function_shift(src, shift)
+def shift(src: DataAndMetadata.DataAndMetadata, shift: typing.Tuple[float, ...], *, order: int=1) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_shift(src, shift, order=order)
+
+def fourier_shift(src: DataAndMetadata.DataAndMetadata, shift: typing.Tuple[float, ...]) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_fourier_shift(src, shift)
 
 def align(src: DataAndMetadata.DataAndMetadata, target: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, bounds: typing.Union[Core.NormRectangleType, Core.NormIntervalType] = None) -> DataAndMetadata.DataAndMetadata:
     return Core.function_align(src, target, upsample_factor, bounds=bounds)
+
+def fourier_align(src: DataAndMetadata.DataAndMetadata, target: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, bounds: typing.Union[Core.NormRectangleType, Core.NormIntervalType] = None) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_fourier_align(src, target, upsample_factor, bounds=bounds)
 
 def sequence_register_translation(src: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, subtract_means: bool = True, bounds: typing.Union[Core.NormRectangleType, Core.NormIntervalType] = None) -> DataAndMetadata.DataAndMetadata:
     return Core.function_sequence_register_translation(src, upsample_factor, subtract_means)
@@ -266,6 +272,9 @@ def sequence_squeeze_measurement(data_and_metadata: DataAndMetadata.DataAndMetad
 
 def sequence_align(src: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, bounds: typing.Union[Core.NormRectangleType, Core.NormIntervalType] = None) -> DataAndMetadata.DataAndMetadata:
     return Core.function_sequence_align(src, upsample_factor, bounds=bounds)
+
+def sequence_fourier_align(src: DataAndMetadata.DataAndMetadata, upsample_factor: int = 1, bounds: typing.Union[Core.NormRectangleType, Core.NormIntervalType] = None) -> DataAndMetadata.DataAndMetadata:
+    return Core.function_sequence_fourier_align(src, upsample_factor, bounds=bounds)
 
 def sequence_integrate(src: DataAndMetadata.DataAndMetadata) -> DataAndMetadata.DataAndMetadata:
     return Core.function_sequence_integrate(src)